+++ src/commands/database/mod.rs
@@ -1,0 +1,8 @@
+mod describe;
+mod list;
+mod tables;
+pub mod utils;
+
+pub use describe::describe_table;
+pub use list::list;
+pub use tables::list_tables;